fix: resolve all compiler warnings for a clean -Wall -Wextra build
- process_monitor.c: read_proc_stat() combined assignment-suppression
  with a length modifier (%*ld) in the /proc/[pid]/stat scanf format,
  which gcc flags as non-portable; switch the discarded fields to %*d
  since their values are never read
- gui_process_panel.c: was missing the include for
  gui_process_integration.h, so is_gui_process_scan_in_progress() and
  gui_compatible_scan_processes() were called via implicit
  declarations instead of their real prototypes
- gui_process_integration.h: drop a verbatim duplicate prototype of
  gui_compatible_scan_processes()
- gui_ports_panel.c: on_scan_ports_clicked() takes a GtkButton*, but
  its two direct call sites passed the GtkWidget* button pointer
  without the cast; add GTK_BUTTON(...)
- gui_backend_adapters.c: wrap_text_for_pdf()/count_wrapped_lines()
  compared a size_t line length against an int max_width; cast to
  size_t at the comparison
- gui_process_panel.c, gui_process_integration.c: several log/debug
  message buffers were sized to fit only the fixed text, so a
  worst-case process name could truncate the %s; size them for the
  actual worst case instead
- gui_process_integration.c: fix a mojibake byte in the high-CPU
  alert message (should be the same 🚨 used by every other alert)
- gui_config_dialog.c, gui_process_panel.c: mark GTK callback
  parameters unused where the signature is fixed by the signal but
  the callback doesn't need them, matching the pattern already used
  elsewhere in the file

Verified with a full rebuild under WSL (GTK+3/Cairo/OpenSSL/libudev):
0 warnings, and the binary still starts, initializes all three
monitoring modules, and shuts down cleanly.

+++ src/gui/integration/gui_backend_adapters.c
@@ -909,7 +909,7 @@ void wrap_text_for_pdf(const char *input, char *output, size_t output_size, int 
             // ================================================================
             // CASO 1: LÍNEA CORTA (CABE EN UNA LÍNEA)
             // ================================================================
-            if (line_len <= max_width) {
+            if (line_len <= (size_t)max_width) {
                 // Copiar línea completa sin modificaciones
                 if (output_pos + line_len < output_size - 1) {
                     strncpy(output + output_pos, input + line_start, line_len);
@@ -1028,7 +1028,7 @@ int count_wrapped_lines(const char *text, int max_width) {    if (!text || max_w
             // ================================================================
             // CÁLCULO MATEMÁTICO DE LÍNEAS NECESARIAS
             // ================================================================
-            if (line_len <= max_width) {
+            if (line_len <= (size_t)max_width) {
                 // Línea corta: ocupa exactamente 1 línea
                 line_count++;
             } else {
